Improve error message for API warnings
Display more information if the upload fails because the API returned a
warning. Therefor the warnings tag is parsed if the result is Warning, and
the keys of the tag are shown as error message (e. g. 'badfilename').
Previously, the error message was only 'Warning', which did not help users
to identify the problem.

 - modify MediaWikiApi.upload to give a more specific error message for
   API warnings

Fix #36