Check for new versions after start.
After LrMediaWiki is loaded, GitHub is checked for new versions if the
user enabled that setting. To parse the response of the GitHub API,
a third-party library is used for JSON parsing (JSON.lua). As the GitHub
API query for the latest release only returns non-pre-release versions,
all releases are queried and the first one is considered the latest.

 - create MediaWikiInit that queries the new version and shows a
   message (if necessary)
 - add MediaWikiInit to Info so that it is run after LrMediaWiki is
   loaded
 - add ‘Check version’ setting to the configuration
 - refactor MediaWikiApi and create performHttpRequest as a generic
   HTTP request method
 - add MediaWikiApi.getCurrentPluginVersion to query the GitHub API

Fix #44.