MediaWikiExportServiceProvider: remove warning if heading is not set
Remove the critical warning when the GPS heading used for the generation of
the Location template is nil.  This does not necessaricily indicate an
invalid value, as previously assumed.  Instead it is possible that this
metadata field is just empty.

Affected code:
 - MediaWikiExportServiceProvider:  remove wrong warning, minor changes
 - TranslatedStrings_de:  remove unnecessary message

Follow-up to c722a7d91ac70325277998e2a07d4041705ffc6d, related to #35.