Improvements of export dialog layout
The layout of the export dialog suffers regarding alignment of the right edges of UI elements.. The static field "Path to the api.php file" at right side isn't viewable in full length at different screen resolutions. It has been removed, the semantic value was low. There ist still room of improvement: The width of UI elemenst differ between Windows and macOS. With these changes the problems are reduced but still not resolved in general.